source destination validation parameters
