+++ luminex/__init__.py
@@ -1,7 +1,6 @@
 from .data_standardization import S3DataLoader
 from .data_standardization import S3DataUploader
 
-from .validation import InputValidator
 from .validation import ETLFileValidator
 from .validation import IAMRoleValidator
 from .validation import ETLValidator
