+++ luminex/run_etl.py
@@ -6,8 +6,7 @@ import time
 
 from github import Github
 from github.GithubException import UnknownObjectException
-from validation import InputValidator
-from validation import ETLFileValidator
+#from validation import ETLFileValidator
 from validation import ETLValidator
 
 # get repo root level
@@ -102,7 +101,7 @@ def submit_spark_job(aws_access_key_id, aws_secret_access_key, aws_session_token
     return {'StepId': step_id, 'Status': step_status}
 
 
-def run_etl(emr_cluster_id, pat, num_transformations, transformation_names, source_path, destination_path):
+def run_etl(emr_cluster_id, pat, num_transformations, transformation_names, source_path, destination_bucket):
     """
     Main function that triggers required functions in the required order to run the transformation on the EMR Cluster.
 
@@ -125,19 +124,8 @@ def run_etl(emr_cluster_id, pat, num_transformations, transformation_names, sour
         return
 
     #ETL Validations
-   # Extract arguments from the command-line
-    source_path = sys.argv[5]
-    destination_path = sys.argv[6]
-    # Example usage
-    validator = ETLValidator(source_path, destination_path)
+    validator = ETLValidator(source_path, destination_bucket)
     validator.run_validation()
-    # # Get the absolute path to the parent directory of the script
-    # script_dir = os.path.dirname(os.path.abspath(__file__))
-    # # Specify the path to the config.json file relative to the script's directory
-    # config_path = os.path.join(script_dir, 'validation', 'config.json')
-    # Create an instance of the InputValidator class and Run the Input validation
-    # input_validator = InputValidator(cfg)
-    # input_validator.run_validation()
     # # Create an instance of the ETLFileValidator class and Run the ETL logic validation
     # etl_validator = ETLFileValidator(cfg)
     # etl_validator.validate_files()
