+++ validateimports.py
@@ -3,5 +3,5 @@ Test that imports do not throw errors.
 """
 
 from luminex import ETLFileValidator
-from luminex import InputValidator
 from luminex import IAMRoleValidator
+from luminex import ETLValidator
